+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 - Apply proxy changes without the need to toggle off/on when changing configurations while proxy is on.
 - Proxy DNS queries when using SOCKS.
+- Add option to toggle proxy on/off with one click (activated in the right-click menu).
 
 ## [1.1.1] 09-18-2017
 ### Fixed